factoring: 마이페이지 불필요한 state제거 및 코드 가시성 개선
