--- src/pages/mypage/type/ImageBox.ts
@@ -1,6 +1,0 @@
-import { UserInfoState } from "@/types/store";
-
-export interface userImageDataProps {
-	userImageData: UserInfoState["photoURL"];
-	$isEditing: boolean;
-}
