+++ src/pages/mypage/type/UserInfoBox.ts
@@ -2,6 +2,4 @@ import { UserInfoState } from "@/types/store";
 
 export interface UserDataProps {
 	userData: UserInfoState;
-	$isEditing: boolean;
-	$setIsEditing: React.Dispatch<React.SetStateAction<boolean>>;
 }
